Restructuring gauge
Starting from scratch.  Components are not the way to go if you need to edit/manipulate the DOM.  Also, this water globe from D3, while cool, is simply too much of a lift to rework into a resuable component.  I will build one from scratch. I should have know it would come to this.  It always comes to this...